Move more logic from import helper to WMS/KML component
